first pass in event and registry system
