+++ src/main/java/me/artuto/endless/commands/botadm/BlacklistGuildCmd.java
@@ -23,7 +23,7 @@ import me.artuto.endless.Bot;
 import me.artuto.endless.Const;
 import me.artuto.endless.cmddata.Categories;
 import me.artuto.endless.commands.EndlessCommand;
-import me.artuto.endless.entities.Blacklist;
+import me.artuto.endless.core.entities.Blacklist;
 import me.artuto.endless.utils.ArgsUtils;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
