+++ src/main/java/me/artuto/endless/commands/botadm/BlacklistUserCmd.java
@@ -23,7 +23,7 @@ import me.artuto.endless.Bot;
 import me.artuto.endless.Const;
 import me.artuto.endless.cmddata.Categories;
 import me.artuto.endless.commands.EndlessCommand;
-import me.artuto.endless.entities.Blacklist;
+import me.artuto.endless.core.entities.Blacklist;
 import me.artuto.endless.utils.ArgsUtils;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
@@ -32,9 +32,7 @@ import net.dv8tion.jda.core.entities.User;
 
 import java.awt.*;
 import java.time.OffsetDateTime;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class BlacklistUserCmd extends EndlessCommand
 {
