+++ src/main/java/me/artuto/endless/commands/utils/TimeForCmd.java
@@ -23,7 +23,7 @@ import com.jagrosh.jdautilities.commons.utils.FinderUtil;
 import me.artuto.endless.Bot;
 import me.artuto.endless.cmddata.Categories;
 import me.artuto.endless.commands.EndlessCommand;
-import me.artuto.endless.entities.Profile;
+import me.artuto.endless.core.entities.Profile;
 import me.artuto.endless.utils.FormatUtil;
 import net.dv8tion.jda.core.entities.Member;
 import net.dv8tion.jda.core.entities.User;
