+++ src/main/java/me/artuto/endless/core/entities/Blacklist.java
@@ -15,10 +15,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.artuto.endless.entities.impl;
+package me.artuto.endless.core.entities;
 
 import me.artuto.endless.Const;
-import me.artuto.endless.entities.Blacklist;
 
 import java.time.OffsetDateTime;
 
@@ -26,42 +25,13 @@ import java.time.OffsetDateTime;
  * @author Artuto
  */
 
-public class BlacklistImpl implements Blacklist
+public interface Blacklist
 {
-    private Const.BlacklistType type;
-    private long id;
-    private OffsetDateTime time;
-    private String reason;
+    Const.BlacklistType getType();
 
-    public BlacklistImpl(Const.BlacklistType type, long id, OffsetDateTime time, String reason)
-    {
-        this.type = type;
-        this.id = id;
-        this.time = time;
-        this.reason = reason;
-    }
+    long getId();
 
-    @Override
-    public Const.BlacklistType getType()
-    {
-        return type;
-    }
+    OffsetDateTime getTime();
 
-    @Override
-    public long getId()
-    {
-        return id;
-    }
-
-    @Override
-    public OffsetDateTime getTime()
-    {
-        return time;
-    }
-
-    @Override
-    public String getReason()
-    {
-        return reason;
-    }
+    String getReason();
 }
