+++ src/main/java/me/artuto/endless/core/entities/impl/GuildSettingsImpl.java
@@ -15,9 +15,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.artuto.endless.entities.impl;
+package me.artuto.endless.core.entities.impl;
 
-import me.artuto.endless.entities.GuildSettings;
+import me.artuto.endless.core.entities.GuildSettings;
 import net.dv8tion.jda.core.entities.Role;
 
 import java.util.Collection;
