+++ src/main/java/me/artuto/endless/handlers/BlacklistHandler.java
@@ -19,16 +19,14 @@ package me.artuto.endless.handlers;
 
 import com.jagrosh.jdautilities.command.CommandEvent;
 import me.artuto.endless.data.managers.BlacklistDataManager;
-import me.artuto.endless.entities.Blacklist;
+import me.artuto.endless.core.entities.Blacklist;
 import net.dv8tion.jda.core.EmbedBuilder;
-import net.dv8tion.jda.core.MessageBuilder;
 import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.awt.*;
-import java.time.format.DateTimeFormatter;
 
 /**
  * @author Artuto
