+++ src/main/java/me/artuto/endless/logging/ServerLogging.java
@@ -23,7 +23,7 @@ import com.jagrosh.jagtag.libraries.*;
 import me.artuto.endless.Bot;
 import me.artuto.endless.Messages;
 import me.artuto.endless.data.managers.GuildSettingsDataManager;
-import me.artuto.endless.entities.ParsedAuditLog;
+import me.artuto.endless.core.entities.ParsedAuditLog;
 import me.artuto.endless.tempdata.MessagesLogging;
 import me.artuto.endless.tools.Variables;
 import me.artuto.endless.utils.*;
@@ -42,7 +42,6 @@ import net.dv8tion.jda.core.events.message.guild.GuildMessageDeleteEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageUpdateEvent;
 import net.dv8tion.jda.core.events.user.update.UserUpdateAvatarEvent;
-import org.apache.commons.lang3.StringUtils;
 
 import java.awt.*;
 import java.io.File;
@@ -174,12 +173,12 @@ public class ServerLogging
                 for(Message.Attachment att : newmsg.getAttachments())
                     newContent.append(att.getUrl()).append("\n");
 
-                String title = "`"+TimeUtils.getTimeAndDate()+" [Message Edited]:` :pencil2: **"+message.getAuthor().getName()+"#"+message.getAuthor().getDiscriminator()+"**'s message was edited in "+message.getTextChannel().getAsMention()+":";
-                String diff = StringUtils.difference(oldContent.toString(), newContent.toString());
-                String diffF ="["+diff+"](https://google.com)";
+                String title = "`"+TimeUtils.getTimeAndDate()+" [Message Edited]:` :pencil2: **"
+                        +message.getAuthor().getName()+"#"+message.getAuthor().getDiscriminator()+"**'s message was edited in "
+                        +message.getTextChannel().getAsMention()+":";
 
                 builder.addField("From:", oldContent.toString(), false);
-                builder.addField("To:", newContent.toString().replace(diff, diffF), false);
+                builder.addField("To:", newContent.toString(), false);
                 builder.setFooter("Message ID: "+message.getId(), null);
                 builder.setColor(Color.YELLOW);
